requirements.txt
contourpy==1.0.7
cycler==0.11.0
darkdetect==0.7.1
fonttools==4.39.3
imageio==2.28.0
imgaug==0.4.0
keras==2.12.0
kiwisolver==1.4.4
lazy-loader==0.2
matplotlib==3.7.1
mtcnn==0.1.1
networkx==3.1
numpy~=1.23.5
opencv-python-headless==4.7.0.72
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pip==23.1.2
pyparsing==3.0.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyQt5==5.15.9
pyqtdarktheme==2.1.0
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
scikit-image==0.20.0
scipy==1.10.1
setuptools==65.5.1
shapely==2.0.1
six==1.16.0
tifffile==2023.4.12
tqdm==4.65.0
tzdata==2023.3
wheel==0.38.4
facenet_pytorch
tensorflow~=2.12.0
torch~=2.0.0
fastapi~=0.95.1
torchvision~=0.15.1
python-multipart
geocoder
requests~=2.29.0